fix: chrome-devtools
Fix inconsistent otter message type, incorrect context and parent detection

Expose rule-engine in the @o3r/chrome-devool public package to allow import in debug panel.

Do not move the component into the o3r/rules-engine but leave in current package as the chrome devtool cannot rely on imported module.

This unfortunately means that we will have a dependence on o3r/chrome-devtools to reuse the panels