fix: avoid lightsaving switch issue on utils.Date
Italy changed their light saving date time logic several time in the past.

As a result, the offset computation for lightsaving switch prior to 1980 are incorrect.

This result in invalid utils.Date which are serialized to the day before.

Setting utils.Date time to 12:00 avoids this error (12:00 being correctly computed as after the time change).

It also reduces the impact of an incorrect offset.