feature: data sources for functional contents
Adds data sources to functional contents to share lists of choices between properties:
For ENUM or STRING properties that need a list a choices, the choices can be provided
either by the "choices" attribute (as it used to be), or by a data source.
The data sources are defined at the root of the functional content definition.
A property references a data source by providing its name in the widget parameter "dataSource".
If both "dataSource" and "choices" are provided by the property, the "choices" have priority.